relocate unit tests to main test location
